requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pandas
python-dotenv
requests